+++ src/preferences.py
@@ -30,6 +30,7 @@ class ImCompressorPrefsWindow(Gtk.Window):
 
     grid = Gtk.Template.Child()
     toggle_new_file = Gtk.Template.Child()
+    entry_suffix = Gtk.Template.Child()
     toggle_dark_theme = Gtk.Template.Child()
 
     _settings = Gio.Settings.new(SETTINGS_SCHEMA)
@@ -46,14 +47,25 @@ class ImCompressorPrefsWindow(Gtk.Window):
         self.toggle_new_file.set_active(self._settings.get_boolean('new-file'))
         self.toggle_new_file.connect('notify::active', self.on_bool_changed,
                                      'new-file')
+
+        # Suffix
+        self.entry_suffix.set_text(self._settings.get_string('suffix'))
+        self.entry_suffix.connect('changed', self.on_string_changed, 'suffix')
+
         # Toggle dark theme
         self.toggle_dark_theme.connect('notify::active', self.on_bool_changed,
                                        'dark-theme')
-        self.toggle_dark_theme.set_active(self._settings.get_boolean('dark-theme'))
+        self.toggle_dark_theme.set_active(
+            self._settings.get_boolean('dark-theme'))
 
     def on_bool_changed(self, switch, state, key):
         self._settings.set_boolean(key, switch.get_active())
         if key == 'dark-theme':
             self.parent.toggle_dark_theme(switch.get_active())
         elif key == 'new-file':
+            self.parent.change_treeview_label()
+
+    def on_string_changed(self, entry, key):
+        self._settings.set_string(key, entry.get_text())
+        if key == 'suffix':
             self.parent.change_treeview_label()